backend/pyproject.toml
[project]
name = "team-architect-backend"
version = "1.0.0"
description = "Team Architect Assistant - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "openai>=1.50.0",
    "gitpython>=3.1.43",
    "msal>=1.31.0",
    "pydantic-settings>=2.5.0",
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.21.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pip-audit>=2.7.0",
    "ruff>=0.6.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
